Hotfix/remove humble servers (#7)
* * temp replace humble node+indexer with algoexplorerapi
* * code cleanup
* * docs css cleanup
* * docs css cleanup (2)
