+++ cli/runAnnouncerTest.mjs
@@ -3,7 +3,7 @@ import { exitWithMsgs, Blue, Red, Yellow, iout, Green } from "./utils.mjs";
 
 let exitTimeout;
 const LIMIT = 10;
-const TIMEOUT = 10;
+const TIMEOUT = 15;
 const pools = new Set();
 
 /** Attach to pool announcer and list a subset of pools */
@@ -40,7 +40,7 @@ function resetTimer() {
 }
 
 /** End CLI */
-export function stopTest() {
+function stopTest() {
   clearTimeout(exitTimeout);
   exitWithMsgs("Timer stopped. Exiting ...");
 }
