+++ cli/runFetchTokenTest.mjs
@@ -1,6 +1,5 @@
 import { fetchToken } from "../lib/index.js";
 import { exitWithMsgs, Blue, Yellow, fromArgs, iout } from "./utils.mjs";
-import { stopTest } from "./runAnnouncerTest.mjs";
 
 /** Fetch and display data for a single token */
 export async function runFetchTokenTest(acc, tokenId) {
@@ -9,5 +8,5 @@ export async function runFetchTokenTest(acc, tokenId) {
 
   Yellow(`Fetching Token "${tokenId}"...`);
   iout("Fetched token", await fetchToken(acc, tokenId));
-  stopTest();
+  exitWithMsgs("Test complete! Exiting ...");
 }
